Add files via upload
This is the BOTDA fiber sensor data that is measured by the System on Chip (SoC) platform. The BOTDA data.zip directory contains 201 .txt files each file containing 300 samples. In addition, there is the SoC_data.txt file containing all the data in the BOTDA data.zip directory but in matrix format, which can be used in the SoC for online processing.